requirements.txt
openai>=1.25.0
python-dotenv>=1.1.1
tavily-python>=0.7.10
langgraph>=0.6.2
tiktoken
httpx
langsmith
fastapi
uvicorn[standard]
python-docx
fpdf
jinja2
